refactor(register page): move folder register to folder pages and rewrite import path
